pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "encrypted-data-manager"
version = "0.1.0"
description = "Local encrypted data query and management service"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn>=0.35.0,<1.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "pyjwt>=2.10.0,<3.0.0",
  "argon2-cffi>=25.1.0,<26.0.0",
  "bcrypt>=4.1.0,<5.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "openpyxl>=3.1.5,<4.0.0",
  "cryptography>=46.0.0,<47.0.0",
  "pandas>=2.3.3",
  "polars>=1.30.0",
  "fastexcel>=0.16.0",
]

[project.optional-dependencies]
test = [
  "pytest>=9.0.0,<10.0.0",
  "httpx>=0.28.0,<1.0.0"
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-p no:capture"
